Fix:모달창 분리, 함수 분리 - LogInOutDialog.tsx    로그인, 로그아웃 모달창을 따로 위의 파일에 분리하였습니다 - user.ts   로그인, 로그아웃, 내 정보를 호출하는 API를 따로 분리하였습니다 - NaviagionBar.tsx   위의 내용들을 분리 후 일반 컴포넌트 역할만 할 수 있게 수정하였습니다
